src/test/java/com/hiep/finalproject/dao/CampaignDAOTest.java
package com.hiep.finalproject.dao;

import com.hiep.finalproject.entity.Campaign;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.context.junit4.SpringRunner;

import javax.persistence.EntityManager;
import javax.sql.DataSource;
import java.util.Optional;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
import static org.junit.jupiter.api.Assertions.*;
@ExtendWith(SpringExtension.class)
@DataJpaTest
class CampaignDAOTest {
    @Autowired  DataSource dataSource;
    @Autowired  EntityManager entityManager;
    @Autowired  CampaignDAO campaignDAO;

    @Test
    void findCampaign() {
        assertThat(dataSource).isNotNull();
        assertThat(entityManager).isNotNull();
        assertThat(campaignDAO).isNotNull();
    }
}